ref: move some flame files around
