+++ app/(app)/components/flames/constants/levels.ts
