+++ app/(app)/components/flames/constants/state.ts
