+++ app/(app)/components/flames/flame-renderer.tsx
@@ -5,7 +5,7 @@ import { motion, useReducedMotion } from 'framer-motion';
 import { ShakeWrapper } from '../../flames/components/flame-card/effects/ShakeWrapper';
 import type { ShapeColors } from '../../flames/components/flame-card/effects/types';
 import { FLAME_REGISTRY } from '../../flames/components/flame-card/flames';
-import type { FlameState } from '../../flames/utils/types';
+import type { FlameState } from './constants/state';
 
 const stateVariants: Record<FlameState, TargetAndTransition> = {
   untended: { scale: 0.8, opacity: 0.88, y: 0 },
