+++ app/(app)/flames/components/flame-card/FlameCard.tsx
@@ -4,12 +4,12 @@ import { motion, useReducedMotion } from 'framer-motion';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import type { Flame } from '@/lib/supabase/rows';
 import { cn } from '@/lib/utils';
+import { getFlameColors } from '../../../components/flames/constants/colors';
+import { getFlameLevel } from '../../../components/flames/constants/levels';
+import type { FlameState } from '../../../components/flames/constants/state';
 import { FlameRenderer } from '../../../components/flames/flame-renderer';
 import { useFlameInteractions } from '../../hooks/useFlameInteractions';
 import type { FlameCardActions, FlameEntry } from '../../hooks/useFlames';
-import { getFlameColors } from '../../utils/colors';
-import { getFlameLevel } from '../../utils/levels';
-import type { FlameState } from '../../utils/types';
 import { CompletionSummaryModal } from '../CompletionSummaryModal';
 import { CompletionCelebration } from './effects/CompletionCelebration';
 import { CompletionRingProgress } from './effects/CompletionRingProgress';
