+++ app/(app)/flames/components/flame-card/InteractiveFlameCard.tsx
@@ -4,8 +4,8 @@ import { useTranslations } from 'next-intl';
 import { toast } from 'sonner';
 import type { Flame } from '@/lib/supabase/rows';
 import { cn } from '@/lib/utils';
+import { getFlameColors } from '../../../components/flames/constants/colors';
 import type { FlameCardActions, FlameEntry } from '../../hooks/useFlames';
-import { getFlameColors } from '../../utils/colors';
 import { FlameCard } from './FlameCard';
 import { ProgressBar } from './ProgressBar';
 import { TimerDisplay } from './TimerDisplay';
