+++ app/(app)/flames/schedule/components/DayRowFuelBar.tsx
@@ -1,11 +1,11 @@
 'use client';
 
 import { useMemo } from 'react';
-import type { FlameColorName } from '@/app/(app)/flames/utils/colors';
+import type { FlameColorName } from '@/app/(app)/components/flames/constants/colors';
 import {
   FLAME_HEX_COLORS,
   getFlameColors,
-} from '@/app/(app)/flames/utils/colors';
+} from '@/app/(app)/components/flames/constants/colors';
 import type { Flame } from '@/lib/supabase/rows';
 import { cn } from '@/lib/utils';
