+++ app/(marketing)/components/FlameShowcase.tsx
@@ -3,11 +3,11 @@
 import { motion, useInView, useReducedMotion } from 'framer-motion';
 import { useTranslations } from 'next-intl';
 import { useRef } from 'react';
+import { FLAME_HEX_COLORS } from '@/app/(app)/components/flames/constants/colors';
+import { FLAME_LEVELS } from '@/app/(app)/components/flames/constants/levels';
 import { FlameRenderer } from '@/app/(app)/components/flames/flame-renderer';
 import { EffectsRenderer } from '@/app/(app)/flames/components/flame-card/effects/EffectsRenderer';
 import { FLAME_REGISTRY } from '@/app/(app)/flames/components/flame-card/flames';
-import { FLAME_HEX_COLORS } from '@/app/(app)/flames/utils/colors';
-import { FLAME_LEVELS } from '@/app/(app)/flames/utils/levels';
 import { ShowcaseFuelBar } from './ShowcaseFuelBar';
 
 const EASE_OUT_EXPO = [0.21, 0.47, 0.32, 0.98] as const;
